+++ 30SecHero/30SecHero/Assets/Scripts/Data/WeaponData.cs
@@ -60,7 +60,7 @@ public class WeaponData : EquipData
         int randIndex = UnityEngine.Random.Range(0, keys.Count);
         return keys[randIndex];
     }
-    public override int SetUID()
+    public override long SetUID()
     {
         base.SetUID();
         MaxUID++;
